Add sentry_flutter

+++ lib/UI/Screens/trips/components/delete_trip.dart
@@ -1,35 +1,38 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:sharqia_household_survey/UI/Screens/trips/provider/trip_provider.dart';
-
 import '../../../../Resources/colors.dart';
 import 'text_trip.dart';
 
 class DeleteTripComponent extends StatelessWidget {
   final int i;
 
-  const DeleteTripComponent({super.key, required this.i});
+  const DeleteTripComponent({required this.i, Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    // TODO: implement build
     final validationService = Provider.of<TripProvider>(context, listen: false);
+    final canDelete = i >= 1;
+
     return Row(
       children: [
         const Spacer(),
-        TextTrip(index: i),
+        TextTrip(
+          index: i,
+        ),
         const Spacer(),
-        i >= 1
-            ? IconButton(
-                onPressed: () {
-                  validationService.removeTrip(i);
-                },
-                icon: Icon(
-                  Icons.delete,
-                  color: ColorManager.primaryColor,
-                ),
-              )
-            : Container()
+        if (canDelete)
+          GestureDetector(
+            onTap: () {
+              validationService.removeTrip(i);
+            },
+            child: Icon(
+              Icons.delete,
+              color: ColorManager.primaryColor,
+            ),
+          )
+        else
+          const SizedBox(),
       ],
     );
   }
